auth-server: setup: Factor setup helpers into separate module
